ENH: modificat el tipus de dades de handleEvent: long enlloc de int

+++ starviewer/src/core/representationslayer.cpp
@@ -129,7 +129,7 @@ void RepresentationsLayer::clearAll()
     }
 }
 
-void RepresentationsLayer::handleEvent(unsigned int eventID, double posX, double posY)
+void RepresentationsLayer::handleEvent(long unsigned eventID, double posX, double posY)
 {
     //De moment no fa res...
 }
